CCS-457: Move str_to_bool helper func to utils
